Exercise4

Description

Appetizer

data Student

Student record.

Constructors

Student

Fields

fname :: String

First name.

lname :: String

Last name.

snum :: Int

Student number.

f2 :: IO Student

Reads a Student from stdin. This is a straighforward translation of the f2 function from Clean to Haskell. The only thing I couldn't get right is catching the exception. I'm quite new to them, and from the implementation of <|> for IO:

instance Alternative IO where
    empty = failIO "mzero"
    (<|>) = mplusIO

-- with
mplusIO m n = m `catchException` \ (_ :: IOError) -> n

I expected <|> to catch the exception thrown by read.

f3 :: IO Student

Reads a Student from stdin. One it's very similar to f2, even the implementation looks pretty much the same.

Serialisation using Functor, Applicative and Monad

newtype Serialised

According to the exercise, I'm supposed to come up with a type to hold serialised version of objects that supports amortised \(\mathcal{O}(1)\) reading and writing.

I've actually made a mistake here as appending to list is an \(\mathcal{O}(N)\) operation. Don't know why I haven't noticed it immediately, but it's quite a bit of work do re-implement everything, so I'll leave it be.

Constructors

Serialised [String]

State access for primitive types

newtype State s a

Our own definition of State monad so that we can reimplement ($), (>>=), (*) etc. This is my second try at this. I've first implemented everything for State as s -> (a, s). For Alternative, however, Maybe does come in handy. So I thought, maybe a MaybeT monad transformer will help. Turned out I needed 'StateT'+'Maybe' combination rather than 'MaybeT'+'State', but StateT is a bit too off topic, so I've decided to just use s -> (Maybe a, s) :D

Constructors

State (s -> (Maybe a, s))

type Serialise a = State Serialised a

A type from the exercise.

rd :: Serialise String

Returns the next string in the state.

wrt :: Show a => a -> Serialise String

Serialises the argument and adds it to the state.

Monad for State

See State.

Matching strings

match :: Show a => a -> Serialise a

match x Compares the serialised version of x to the next string in the state. If these match, Just x is returned, otherwise Nothing.

pred :: (String -> Bool) -> Serialise String

Very similar to match except that a predicate is used rather than operator (==).

Serialisation of Generic Types

In the previous exercise I made extensive use of the ReadP monad, i.e. all was already done using Monad, Applicative, and Alternative classes.

Anyways, to make better use of the tools developed in this exercise, let's adapt our Serialise class a bit:

class Serialisable a where

Out serialisation. Notice how we now use Serialise rather than ReadP and ShowS.

Methods

readPrec' :: Int -> Serialise a

writePrec' :: Int -> a -> Serialise String

readPrec' :: (Generic a, SerialiseImpl (Rep a)) => Int -> Serialise a

writePrec' :: (Generic a, SerialiseImpl (Rep a)) => Int -> a -> Serialise String

data Bin a

Simple tree.

Constructors

Leaf

Bin (Bin a) a (Bin a)

read' :: Serialisable a => Serialised -> Maybe a

Our analogue of read

show' :: Serialisable a => a -> Serialised

Out analogue of show